handle paths with colons
As it turns out, IRequestInfo.getHeaders() returns a list of strings
that happens to include the first line of the request as well (HTTP
verb, path and version). This was usually caught by the check that
ignored entries without a colon, however, if the path itself contained
one, the first line was passed into the list of headers as well,
resulting in an invalid artificial header.

This commit fixes the issue (reported by @mirusu400 in #3) by explicitly
ignoring the first entry of the list returned by getHeaders().